Vim plugin checks and location list
The JSON formatter supports linewize and visualwise Vim ranges allowing the user to choose what to format.

A separate Vim window / buffer is opened to display messages and error messages.  The location and size of that window is configurable.

Better error and reporting is displayed after a format of the json from the node.js module (changes on that side are also required).

Changes integrated into the Vim location list, so that you can click directly on a line and it will take you to where the error is.  This is the same features of Vim used for the :grep and :make commands.